CLang Sanitizer

+++ tests/flow/tests.sh
@@ -3,12 +3,16 @@
 # [[ $VERBOSE == 1 ]] && set -x
 [[ $IGNERR == 1 ]] || set -e
 
-HERE="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null 2>&1 && pwd)"
+PROGNAME="${BASH_SOURCE[0]}"
+HERE="$(cd "$(dirname "$PROGNAME")" &>/dev/null && pwd)"
 ROOT=$(cd $HERE/../.. && pwd)
-READIES=$ROOT/deps/readies 
+READIES=$ROOT/deps/readies
 . $READIES/shibumi/defs
 
 VALGRIND_REDIS_VER=6.2
+SAN_REDIS_VER=6.2
+
+cd $HERE
 
 #----------------------------------------------------------------------------------------------
 
@@ -30,33 +34,118 @@ help() {
 		RLEC=0|1            General tests on RLEC
 
 		REDIS_SERVER=path   Location of redis-server
-		EXT|EXISTING_ENV=1  Run the tests on existing env
-
+		EXISTING_ENV=1      Run the tests on existing env
+		EXT=1|run           Test on existing env (1=running; run=start redis-server)
+		EXT_HOST=addr       Address if existing env (default: 127.0.0.1)
+		EXT_PORT=n          Port of existing env
 		TEST=test           Run specific test (e.g. test.py:test_name)
-		VALGRIND|VG=1       Run with Valgrind
+		COV=1               Run with coverage analysis
+		VG=1                Run with Valgrind
+		VG_LEAKS=0          Do not detect leaks
+		SAN=type            Use LLVM sanitizer (type=address|memory|leak|thread) 
+		GDB=1               Enable interactive gdb debugging (in single-test mode)
 
 		DOCKER_HOST         Address of Docker server (default: localhost)
 		RLEC_PORT           Port of existing-env in RLEC container (default: 12000)
 
 		RLTEST_ARGS=...     Extra RLTest arguments
 		VERBOSE=1           Print commands
+		LOG=1               Send results to log (even on single-test mode)
+		KEEP=1              Do not remove intermediate files
 		IGNERR=1            Do not abort on error
+		NOP=1               Dry run
+		HELP=1              Show help
 
 	END
+	exit 0
+}
+
+#---------------------------------------------------------------------------------------------- 
+
+setup_rltest() {
+	if [[ $RLTEST == view ]]; then
+		if [[ ! -d $ROOT/../RLTest ]]; then
+			eprint "RLTest not found in view $ROOT"
+			exit 1
+		fi
+		RLTEST=$(cd $ROOT/../RLTest; pwd)
+	fi
+
+	if [[ -n $RLTEST ]]; then
+		if [[ ! -d $RLTEST ]]; then
+			eprint "Invalid RLTest location: $RLTEST"
+			exit 1
+		fi
+
+		# Specifically search for it in the specified location
+		export PYTHONPATH="$PYTHONPATH:$RLTEST"
+		if [[ $VERBOSE == 1 ]]; then
+			echo "PYTHONPATH=$PYTHONPATH"
+		fi
+	fi
+
+	if [[ -n $TEST ]]; then
+		RLTEST_ARGS+=" --test $TEST"
+		export BB=${BB:-1}
+	fi
+
+	GDB=${GDB:-0}
+	if [[ $GDB == 1 ]]; then
+		RLTEST_ARGS+=" -i --verbose"
+	fi
+
+	if [[ $PARALLEL == 1 ]]; then
+		RLTEST_PARALLEL_ARG="--parallelism $($READIES/bin/nproc)"
+	fi
+}
+
+#----------------------------------------------------------------------------------------------
+
+setup_clang_sanitizer() {
+	if ! grep THPIsEnabled /build/redis.blacklist &> /dev/null; then
+		echo "fun:THPIsEnabled" >> /build/redis.blacklist
+	fi
+
+
+	# for RLTest
+	export SANITIZER="$SAN"
+	
+	# --no-output-catch --exit-on-failure --check-exitcode
+	RLTEST_SAN_ARGS="--unix --sanitizer $SAN"
+
+	if [[ $SAN == addr || $SAN == address ]]; then
+		REDIS_SERVER=${REDIS_SERVER:-redis-server-asan-$SAN_REDIS_VER}
+		if ! command -v $REDIS_SERVER > /dev/null; then
+			echo Building Redis for clang-asan ...
+			$READIES/bin/getredis --force -v $SAN_REDIS_VER --own-openssl --no-run --suffix asan --clang-asan --clang-san-blacklist /build/redis.blacklist
+		fi
+
+		export ASAN_OPTIONS=detect_odr_violation=0
+		# :detect_leaks=0
+
+	elif [[ $SAN == mem || $SAN == memory ]]; then
+		REDIS_SERVER=${REDIS_SERVER:-redis-server-msan-$SAN_REDIS_VER}
+		if ! command -v $REDIS_SERVER > /dev/null; then
+			echo Building Redis for clang-msan ...
+			$READIES/bin/getredis --force -v $SAN_REDIS_VER  --no-run --own-openssl --suffix msan --clang-msan --llvm-dir /opt/llvm-project/build-msan --clang-san-blacklist /build/redis.blacklist
+		fi
+	fi
+}
+
+clang_sanitizer_summary() {
+	if grep -l "leaked in" logs/*.asan.log* &> /dev/null; then
+		echo
+		echo "${LIGHTRED}Sanitizer: leaks detected:${RED}"
+		grep -l "leaked in" logs/*.asan.log*
+		echo "${NOCOLOR}"
+		E=1
+	fi
 }
 
 #----------------------------------------------------------------------------------------------
 
 setup_redis_server() {
-	if [[ $VALGRIND == 1 ]]; then
-		REDIS_SERVER=${REDIS_SERVER:-redis-server-vg}
-		if ! is_command $REDIS_SERVER; then
-			echo Building Redis for Valgrind ...
-			$READIES/bin/getredis -v $VALGRIND_REDIS_VER --valgrind --suffix vg
-		fi
-	else
-		REDIS_SERVER=${REDIS_SERVER:-redis-server}
-	fi
+	REDIS_SERVER=${REDIS_SERVER:-redis-server}
 
 	if ! is_command $REDIS_SERVER; then
 		echo "Cannot find $REDIS_SERVER. Aborting."
@@ -66,8 +155,16 @@ setup_redis_server() {
 
 #----------------------------------------------------------------------------------------------
 
+setup_valgrind() {
+	REDIS_SERVER=${REDIS_SERVER:-redis-server-vg}
+	if ! is_command $REDIS_SERVER; then
+		echo Building Redis for Valgrind ...
+		$READIES/bin/getredis -v $VALGRIND_REDIS_VER --valgrind --suffix vg
+	fi
+}
+
 valgrind_config() {
-	export VG_OPTIONS="
+	export VG_OPTIONS="\
 		-q \
 		--leak-check=full \
 		--show-reachable=no \
@@ -79,7 +176,7 @@ valgrind_config() {
 
 	VALGRIND_SUPRESSIONS=$ROOT/tests/redis_valgrind.sup
 
-	VALGRIND_ARGS+="\
+	RLTEST_VALGRIND_ARGS+="\
 		--no-output-catch \
 		--use-valgrind \
 		--vg-verbose \
@@ -88,10 +185,33 @@ valgrind_config() {
 	export VALGRIND=1
 }
 
+valgrind_summary() {
+	# Collect name of each flow log that contains leaks
+	FILES_WITH_LEAKS=$(grep -l "definitely lost" logs/*.valgrind.log)
+	if [[ ! -z $FILES_WITH_LEAKS ]]; then
+		echo "Memory leaks introduced in flow tests."
+		echo $FILES_WITH_LEAKS
+		# Print the full Valgrind output for each leaking file
+		echo $FILES_WITH_LEAKS | xargs cat
+		exit 1
+	else
+		echo Valgrind test ok
+	fi
+}
+
+#----------------------------------------------------------------------------------------------
+
+setup_coverage() {
+	RLTEST_COV_ARGS="--unix"
+
+	export CODE_COVERAGE=1
+}
+
 #----------------------------------------------------------------------------------------------
 
 run_tests() {
 	local title="$1"
+	shift
 	if [[ -n $title ]]; then
 		$READIES/bin/sep -0
 		printf "Running $title:\n\n"
@@ -107,28 +227,57 @@ run_tests() {
 				--module $MODULE
 				--module-args '$MODARGS'
 				$RLTEST_ARGS
-				$VALGRIND_ARGS
+				$RLTEST_PARALLEL_ARG
+				$RLTEST_VALGRIND_ARGS
+				$RLTEST_SAN_ARGS
+				$RLTEST_COV_ARGS
 
 				EOF
 		else
 			cat <<-EOF > $rltest_config
 				--clear-logs
 				$RLTEST_ARGS
-				$VALGRIND_ARGS
+				$RLTEST_VALGRIND_ARGS
 
 				EOF
 		fi
 	else # existing env
-		rltest_config=$(mktemp "${TMPDIR:-/tmp}/xredis_rltest.XXXXXXX")
-		rm -f $rltest_config
-		cat <<-EOF > $rltest_config
-			--env existing-env
-			$RLTEST_ARGS
+		if [[ $EXT == run ]]; then
+			xredis_conf=$(mktemp "${TMPDIR:-/tmp}/xredis_conf.XXXXXXX")
+			rm -f $xredis_conf
+			cat <<-EOF > $xredis_conf
+				loadmodule $MODULE $MODARGS
+				EOF
 
-			EOF
+			rltest_config=$(mktemp "${TMPDIR:-/tmp}/xredis_rltest.XXXXXXX")
+			rm -f $rltest_config
+			cat <<-EOF > $rltest_config
+				--env existing-env
+				$RLTEST_ARGS
+
+				EOF
+
+			if [[ $VERBOSE == 1 ]]; then
+				echo "External redis-server configuration:"
+				cat $xredis_conf
+			fi
+
+			$REDIS_SERVER $xredis_conf &
+			XREDIS_PID=$!
+			echo "External redis-server pid: " $XREDIS_PID
+
+		else # EXT=1
+			rltest_config=$(mktemp "${TMPDIR:-/tmp}/xredis_rltest.XXXXXXX")
+			[[ $KEEP != 1 ]] && rm -f $rltest_config
+			cat <<-EOF > $rltest_config
+				--env existing-env
+				--existing-env-addr $EXT_HOST:$EXT_PORT
+				$RLTEST_ARGS
+
+				EOF
+		fi
 	fi
 
-	cd $ROOT/tests/flow
 
 	if [[ $VERBOSE == 1 ]]; then
 		echo "RLTest configuration:"
@@ -146,15 +295,37 @@ run_tests() {
 
 	[[ $KEEP != 1 ]] && rm -f $rltest_config
 
+	if [[ -n $XREDIS_PID ]]; then
+		echo "killing external redis-server: $XREDIS_PID"
+		kill -TERM $XREDIS_PID
+	fi
+
 	return $E
 }
 
 #----------------------------------------------------------------------------------------------
 
-[[ $1 == --help || $1 == help ]] && {
-	help
-	exit 0
-}
+[[ $1 == --help || $1 == help || $HELP == 1 ]] && { help; exit 0; }
+
+if [[ $RLEC != 1 ]]; then
+	MODULE="${MODULE:-$1}"
+	if [[ -z $MODULE || ! -f $MODULE ]]; then
+		echo "Module not found at ${MODULE}. Aborting."
+		exit 1
+	fi
+fi
+
+if [[ $REDIS_VERBOSE == 1 || $VERBOSE == 1 ]]; then
+	if [[ $LOG != 1 ]]; then
+		RLTEST_ARGS+=" -s -v"
+	fi
+fi
+
+if [[ $COV == 1 ]]; then
+	setup_coverage
+fi
+
+[[ $EXT == 1 || $EXT == run ]] && EXISTING_ENV=1
 
 OP=""
 [[ $NOP == 1 ]] && OP=echo
@@ -167,9 +338,35 @@ AOF=${AOF:-1}
 AOF_SLAVES=${AOF_SLAVES:-1}
 OSS_CLUSTER=${OSS_CLUSTER:-0}
 SHARDS=${SHARDS:-3}
-RLEC=${RLEC:-0}
 
-[[ $EXT == 1 ]] && EXISTING_ENV=1
+[[ $SAN == addr ]] && SAN=address
+[[ $SAN == mem ]] && SAN=memory
+
+EXT_HOST=${EXT_HOST:-127.0.0.1}
+EXT_PORT=${EXT_PORT:-6379}
+
+DOCKER_HOST=${DOCKER_HOST:-localhost}
+
+RLEC=${RLEC:-0}
+RLEC_PORT=${RLEC_PORT:-12000}
+
+[[ $EXT == 1 || $EXT == run || $EXISTING_ENV == 1 ]] && PARALLEL=0
+
+#----------------------------------------------------------------------------------------------
+
+setup_rltest
+
+if [[ -n $SAN ]]; then
+	setup_clang_sanitizer
+fi
+
+if [[ $VG == 1 ]]; then
+	setup_valgrind
+fi
+
+if [[ $RLEC != 1 ]]; then
+	setup_redis_server
+fi
 
 #----------------------------------------------------------------------------------------------
 
@@ -183,40 +380,6 @@ fi
 
 #----------------------------------------------------------------------------------------------
 
-DOCKER_HOST=${DOCKER_HOST:-localhost}
-RLEC_PORT=${RLEC_PORT:-12000}
-
-#----------------------------------------------------------------------------------------------
-
-if [[ $RLEC != 1 ]]; then
-	MODULE=${MODULE:-$1}
-	[[ -z $MODULE || ! -f $MODULE ]] && {
-		echo "Module not found at ${MODULE}. Aborting."
-		exit 1
-	}
-fi
-
-GDB=${GDB:-0}
-
-[[ $VG == 1 ]] && VALGRIND=1
-if [[ $VALGRIND == 1 ]]; then
-	valgrind_config
-fi
-
-if [[ -n $TEST ]]; then
-	RLTEST_ARGS+=" --test $TEST"
-	export BB=${BB:-1}
-fi
-
-[[ $VERBOSE == 1 ]] && RLTEST_ARGS+=" -v"
-[[ $GDB == 1 ]] && RLTEST_ARGS+=" -i --verbose"
-
-cd $ROOT/tests/flow
-
-[[ $RLEC != 1 ]] && setup_redis_server
-
-# RLTEST_ARGS+=" --print-server-cmd"
-
 E=0
 [[ $GEN == 1 ]]    && { (run_tests "general tests"); (( E |= $? )); } || true
 [[ $SLAVES == 1 ]] && { (RLTEST_ARGS="${RLTEST_ARGS} --use-slaves" run_tests "tests with slaves"); (( E |= $? )); } || true
@@ -225,5 +388,9 @@ E=0
 [[ $OSS_CLUSTER == 1 ]] && { (RLTEST_ARGS="${RLTEST_ARGS} --env oss-cluster --shards-count $SHARDS" run_tests "tests on OSS cluster"); (( E |= $? )); } || true
 
 [[ $RLEC == 1 ]]   && { (RLTEST_ARGS="${RLTEST_ARGS} --env existing-env --existing-env-addr $DOCKER_HOST:$RLEC_PORT" run_tests "tests on RLEC"); (( E |= $? )); } || true
+
+if [[ $NOP != 1 && -n $SAN ]]; then
+	clang_sanitizer_summary
+fi
 
 exit $E
